requirements/common.txt
Django<1.8
django-cms<3.1
